tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.relativecolor.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.gradient.relativecolor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.relativecolor</h1>
<p class="desc">Relative color works as CanvasGradient color input</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Relative color works as CanvasGradient color input");
_addTest(function(canvas, ctx) {

  var g = ctx.createLinearGradient(0, 0, 100, 0);
  g.addColorStop(0, '#f00');
  g.addColorStop(1, 'rgb(from red g r b)');
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  _assertPixelApprox(canvas, 25,25, 236,116,0,255, 3);
  _assertPixelApprox(canvas, 50,25, 208,170,0,255, 3);
  _assertPixelApprox(canvas, 75,25, 159,214,0,255, 3);

});
</script>
